requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.1.1
python-dotenv==1.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
